Move expectations so that Bridgetown::Test can access it too
